+++ kernel/arch/x86_64/vmm/vmmapper.c
@@ -13,6 +13,7 @@
 
 #ifdef UNIT_TESTS
 #include "mock_cpu.h"
+#include "mock_recursive.h"
 #else
 #include "x86_64/kdrivers/cpu.h"
 #endif
