+++ kernel/ipc/channel.c
@@ -40,6 +40,7 @@
 
 #ifdef UNIT_TESTS
 #define STATIC_EXCEPT_TESTS
+#include "mock_recursive.h"
 #else
 #define STATIC_EXCEPT_TESTS static
 #endif
