+++ kernel/tests/arch/x86_64/process/address_space_init.c
@@ -12,7 +12,7 @@
 #include "mock_vmm.h"
 #include "vmm/vmmapper.h"
 
-#include "process/address_space.h"
+#include "x86_64/process/address_space.h"
 
 #define TEST_PML4_ADDR (((uint64_t *)0x100000))
 #define TEST_PAGE_COUNT ((32768))
