+++ kernel/tests/fba/alloc.c
@@ -11,10 +11,12 @@
 #include <stdint.h>
 #include <stdlib.h>
 
+#include "munit.h"
+
 #include "fba/alloc.h"
 #include "mock_pmm.h"
+#include "mock_recursive.h"
 #include "mock_vmm.h"
-#include "munit.h"
 #include "structs/bitmap.h"
 #include "vmm/vmmapper.h"
