+++ kernel/tests/include.mk
@@ -9,6 +9,7 @@ CLEAN_ARTIFACTS+=kernel/tests/*.o kernel/tests/pmm/*.o kernel/tests/vmm/*.o 		\
 				kernel/tests/arch/x86_64/process/*.o								\
 				kernel/tests/arch/x86_64/structs/*.o								\
 				kernel/tests/arch/x86_64/vmm/*.o									\
+				kernel/tests/arch/riscv64/*.o										\
 				kernel/tests/*.gcda kernel/tests/pmm/*.gcda kernel/tests/vmm/*.gcda	\
 				kernel/tests/structs/*.gcda kernel/tests/pci/*.gcda 				\
 				kernel/tests/fba/*.gcda kernel/tests/slab/*.gcda 					\
@@ -21,6 +22,7 @@ CLEAN_ARTIFACTS+=kernel/tests/*.o kernel/tests/pmm/*.o kernel/tests/vmm/*.o 		\
 				kernel/tests/arch/x86_64/process/*.gcda								\
 				kernel/tests/arch/x86_64/structs/*.gcda								\
 				kernel/tests/arch/x86_64/vmm/*.gcda									\
+				kernel/tests/arch/riscv64/*.gcda									\
 				kernel/tests/*.gcno kernel/tests/pmm/*.gcno kernel/tests/vmm/*.gcno	\
 				kernel/tests/structs/*.gcno kernel/tests/pci/*.gcno 				\
 				kernel/tests/fba/*.gcno kernel/tests/slab/*.gcno 					\
@@ -34,6 +36,7 @@ CLEAN_ARTIFACTS+=kernel/tests/*.o kernel/tests/pmm/*.o kernel/tests/vmm/*.o 		\
 				kernel/tests/arch/x86_64/structs/*.gcno								\
 				kernel/tests/arch/x86_64/structs/*.gcno								\
 				kernel/tests/arch/x86_64/vmm/*.gcno									\
+				kernel/tests/arch/riscv64/*.gcno									\
 				kernel/tests/build													\
 				gcov/kernel
