+++ kernel/tests/slab/alloc.c
@@ -8,10 +8,12 @@
 #include <stdint.h>
 #include <stdlib.h>
 
+#include "munit.h"
+
 #include "fba/alloc.h"
 #include "mock_pmm.h"
+#include "mock_recursive.h"
 #include "mock_vmm.h"
-#include "munit.h"
 #include "slab/alloc.h"
 #include "vmm/vmconfig.h"
